requirements.txt
requests==2.28
python-dotenv==0.20.0